fix: move backend validation to utils so it can be imported from the front
