+++ src/components/Modal/UpdateProposalStatusModal/UpdateProposalStatusModal.tsx
@@ -11,9 +11,9 @@ import { Modal, ModalProps } from 'decentraland-ui/dist/components/Modal/Modal'
 import Icon from 'semantic-ui-react/dist/commonjs/elements/Icon'
 import isEthereumAddress from 'validator/lib/isEthereumAddress'
 
-import { validateUniqueAddresses } from '../../../back/utils/validations'
 import { Governance } from '../../../clients/Governance'
 import { ProposalAttributes, ProposalStatus, ProposalType } from '../../../entities/Proposal/types'
+import { validateUniqueAddresses } from '../../../entities/Transparency/utils'
 import useFormatMessage from '../../../hooks/useFormatMessage'
 import Label from '../../Common/Typography/Label'
 import Markdown from '../../Common/Typography/Markdown'
