+++ src/entities/Transparency/utils.ts
@@ -54,3 +54,8 @@ export function blockExplorerLink(wallet: TokenInWallet): BlockExplorerLink {
       return { name: '', link: '/' }
   }
 }
+
+export function validateUniqueAddresses(addresses: string[]): boolean {
+  const uniqueSet = new Set(addresses.map((address) => address.toLowerCase()))
+  return uniqueSet.size === addresses.length
+}
